requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
numpy==1.26.4
spacy==3.7.6
sentence-transformers==3.1.1
google-genai==1.11.0
tqdm==4.67.1